Use TorSocket for hidden service self-test
This removes some redundant code and allows better handling of retries
and connectivity state for the hidden service self-test. The self-test
is also run again after connectivity changes to offline.

This should later be used to display some kind of error after a timeout
if the test does not succeed.