ui: Add options to hide the browser nagging dialog
The "open in browser" option on links opens a dialog, meant to warn
users away from using their default browser to open links where they
might want to preserve anonymity.

This commit changes that UI, and adds options to skip that dialog in the
future for links from the specific contact, or for links from all
contacts.