Attempt to remove stale config file locks
QLockFile checks whether a process with the PID in the lock file exists,
and assumes the lock is held if one does. In many cases that can result
in false-positive "locked" errors, especially after unexpected reboots.

It's very bad if we tell users that their configuration is in use when
it isn't. To avoid this, try removing stale lock files automatically. On
most systems, if the lock is truly held, the removal will fail and we
can safely display the error.

It's possible that this could allow users on networked filesystems to
launch multiple instances, but that is less of a bug than the existing
one.

Fixes #73