ui: Fix new chat window opens only to the right #85
On Windows a new chat window got opened always right to the main
windows even if this position was off screen.

To have a consistent behaviour on both single and multi screen systems a
list of all available screens is generated. This list is used to
calculate the relative position of the chat window on the screen where
the main window is located. If the position is off screen the
chat window opens left from the main window.

Also if the main window is moved below the bottom of the screen any
new chat window will be opened above the screen bottom. Note: this could
be problematically on systems with vertically stacked screens.