Add simple version negotiation at the beginning of a connection
This tiny protocol version negotiation is compatible with the Ricochet
1.0 protocol. It's a useful step to have in case there is a need to make
dramatic protocol changes in the future, e.g. with message encryption.
An opt-in feature protocol would be unnecessarily difficult to work with
in some cases.

This also allows us to treat old clients better, and maybe to send them
a useful message so users know that they will need to upgrade.