Send a fake message to contacts using the old protocol to encourage upgrades
Ricochet 1.0.x has no way of indicating protocol issues to users, and
it's not feasible to support both protocols simultaneously, given how
short this upgrade cycle should be.

To try to make that situation slightly less awful, this sends an evil
viral spam message to existing contacts when we connect to them and they
are using the old protocol. That message is sent at most once per
contact, and just encourages them to visit the website for more
information.

I don't like this format, and I don't like encouraging users to follow
links and get updates via messages from their contacts, but I think this
is the best option available: it's better than all of your contacts
appearing to be forever offline after the upgrade, until they happen to
notice the new versions too.