Remove the hidden service self-testing step
It's not necessary to self-test connectivity of the hidden service
immediately after we publish it. This information isn't used for
anything valuable, and in every observed situation, we already have a
good idea of whether our service is connectable or not based on tor's
circuit status.

It has also raised some concerns, because a new Ricochet client with no
contacts will start trying to connect to some scrubbed service in the
logs.

Fixes #26