Unmark 'rejected' contacts if they successfully connect
The 'rejected' mark comes from contacts that reject our authentication,
normally because they removed us as a contact. If that person then sends
a new contact request, it will be automatically accepted, and allowed to
connect. In this case, the rejected mark should be removed.

Also added a BUG for the case where a connection is established with a
contact, and the contact's status doesn't match.