Separate Copy options to clarify message context menu
Add distinct "Copy Message" and "Copy Selection" options in place of the ambiguous "Copy" option

"Copy Message" is available when no text has been selected, or when only a link is selected, such as when a link is left-clicked. Otherwise, "Copy Selection" is available.

This also prevents links from being implicitly selected when left-clicked. So far I haven't found anything that this breaks.

Resolves issue #179.