Making the Initial Setup Nicer
This commit makes some updates to the initial setup UI with the aim of
making it a nicer experience for new users.

![image](https://cloud.githubusercontent.com/assets/106626/8145951/f8befc76-11d2-11e5-98b1-dfbdaa2849fc.png)

The initial setup is the first thing a new user of Ricochet sees when
they start up the software - it should be a place where we a) reassure
them that the software is working b) help them perform intiial setup
which we can't figure out and c) help the user form a connection to the
software so we are better able to present intuitive messages to the
user when we do need their input e.g. if the software discovers
something which needs the users attention like their connection to the
Tor network is blocked.

Presenting complex topics, like proxies and bridges, to a first time
user can be daunting and so the idea behind this initial setup is to
dress up  the initial dialogue and present the user with a sane
default "Press Connect".

There are still issues with this initial setup - in particular when there
is additional configuration needed - we need to refine ways to work with
the user and, where we can, offer suggestions in a safe and secure
manner.

Image copyright information: The images in this commit were created by me
with the exception of the "police officer icon at the center of the
censored icon which is from the Symbola font package
http://users.teilar.gr/~g1951d/ which carries the following licensing terms:
"Fonts and documents in this site are not pieces of property or merchandise
items; they carry no trademark, copyright, license or other market tags;
they are free for any use."