Fix uninitialized read during ContactUser construction
updateStatus is called from the constructor to initialize m_status,
but it compares its value to the previous value of m_status, which
is an uninitialized read. This is harmless in all cases, unless the
value happens to match Offline, which could cause it to not attempt
any outgoing connection.

Found by ubsan.