Use custom scrolling instead of ScrollView
ScrollView is extremely buggy, and seems to regress in every Qt release.
It has a lot of trouble keeping the view scrolled to the bottom, and
with moving at a predictable rate in response to wheel events.

Writing a simple visual scrollbar and wheel handling manually makes for
a much better experience, even if we lose some platform-native scrollbar
behaviors.