Escaping HTML in Contact Users Nicknames
It turns out that if you enter html into a nickname it
doesn't get rendered during the initial contact request, but
if requested, the html will render in the contacts column and
in the setting menu.

Mostly this isn't an issue (the odd style issue when users
set their nicks to be <h1>) however - it is possible, in the
limited 30 character space to construct an <img> tag which -
providing they can create a url short enough is enough to
automatically render the image and reveal the users IP
address.

This very simple fix forces contact nick names to have html escaped.