Prevent Bidi Phishing
Prevents Bidi Phishing attempts by forcing left-to-right display for
links through html entity &#8234;

Adding this as a seperate commit #275 so that the regex
changes can be judged seperately.

This is a fairly minor risk as a victim would have to go through many
hoops and not see the obvious url issues. But better fixed than not.