Fix parsing of links containing escape sequences
When formatting links to be displayed in messages, using repeated calls
to QString::arg() will interpret '%0' and similar in the URL as a
placeholder, resulting in the link not being properly formatted.

Depending on the value, this can either result in a broken URL when
copied to the clipboard (#403), or the URL being displayed with a label
of just '%2' (#372).

This cannot be used to mislabel links, and there is no printf-style
format vulnerability with QString::arg. There is no security impact.