keep app running even if window closed [on OS X]
The default behavior of native OS X apps is to leave the app running
even on closing the last remaining window. Since the
quitOnLastWindowClosed property of QApplication is true by default, we
can remove the line of code that explicitly quits the app on the main
window not being visible.

reference: #343