Remove Esperanto translation due to missing locale
Qt does not have a locale for Esperanto, so the translation does not
behave properly (e.g. not having a name in the languages selection), and
will be incomplete. This is also missing an installer translation, and
is of pretty niche interest.

Removing it for now, until it can meet the usual standards of quality.