ricochet: Run lrelease at qmake time
We cannot set a dependency to force updateqm to run before rcc, which
makes rcc complain (unnecessarily, really) about missing language files.
To prevent this being a source of confusion, switch from using an extra
compiler to running lrelease by hand.

Note that this does mean that updating translations will mean that qmake
will need to be reinvoked as there is no way of specifying the .ts as
build dependencies, but this should not be a source of major burden.